Add safe string from admin
The admin app has some useful tools for turning arbitrary strings into
things that can be used where there are restrictions on character sets
or spacing, such as IDs and the local parts of email addresses.

We’re moving this here so that we can reuse them for turning the names
of regions (e.g. The South West) into IDs that can go into URLs or
database fields (e.g. the-south-west).