Make `BroadcastMessageTemplate` pass schema validation
This commit adds the minimum required properties to the output of
`BroadcastMessageTemplate.__str__` so that it validates against the CAP
XML schema.

At the moment it hard codes a lot of values with what seem like sensible
defaults. We’ll probably need to change these or make them configurable
at some point in the future.