add from_event method and optional msg_type param
we store an event in the database for each action that happens to a
broadcast - so for example, the initial alert, an update, and then a
cancel is three events. For generating the XML we'll move away from
templates and placeholders and towards this.

Note: we'll still need the regular constructor for now to avoid a
chicken/egg loop - but in the future I predict that we'll have a
BroadcastTemplate class that just stores the content and returns
personalised markdown etc, ready for exporting to an event, and then a
BroadcastEventXMLTemplate that takes in an event and returns the XML to
send to the provider