Limit broadcasts to 24 hours
If a broadcast is longer than 24 hours phones that have already
acknowledged the broadcast will repeat it.

This commit limits the time a broadcast can be live for to 23 hours 59
minutes (I added a minute of padding to help avoid edge cases).

This still doesn’t mean that the end time is configurable, but we might
be removing that from the frontend anyway.