Do not break gov.uk if a '=' character precedes it
Currently, if we have 'gov.uk' that does not look like a URL then we add
a small non breaking space between gov. and uk so that it is not
formatted as a link.

This commit is intended to fix the
https://www.pivotaltracker.com/story/show/174510633 ticket where
`gov.uk/?utm_source=gov.uk` would have the second use of 'gov.uk' split
by a non breaking space.

However, in fixing this we decided to change the approach of the regex:

We currently say "lets try and find gov.uk where it's not part of a URL"

Find all usages of "GOV.UK", unless they're preceeded by one of .,/,=

Instead we could flip the logic. "Lets try and find gov.uk where it's presentational

Find all usages of "GOV.UK" where it has whitespace before the GOV.UK.

We think it's better to miss de-linking a niche presentational gov.uk usage
than accidentally de-link a niche gov.uk usage in a url.

We add some extra test cases to support this change that we think could
be interesting edge cases. We also split the test up so that the regex
is tested individually to testing that we do the escaping in templates,
which is now done as more of a single integration test.