Change fields required to create BroadcastMessageTemplate
Previously it was expecting in the BroadcastMessageEvent to have
`polygons` and `areas` as keys. This is not the case currently and we
change utils to match what the API currently does which is instead
provide this information using `transmitted_areas` which is a dict
containing `areas` and `simply_polygons`.

Whilst we do this, we also update the test cases to match the keys of
the test events to the schema of the BroadcastMessageEvent serialize
method.

We also add in an additional test for multiple polygons whilst we are
here too.