Add semantically marked-up heading to broadcast preview
At the moment we add the words ‘Emergency alert’ to the preview of a
broadcast message using CSS generated content.

This isn’t accessible because screen readers ignore generated content.

This commit adds the same content wrapped in an appropriate HTML
element, which can be styled the same but will also be read out by
screenreaders.