Reduce max plain text email content limit to 1MB
The message we send to SES is structured as follows:
"Message": {
  'Subject': {
    'Data': subject,
   },
  'Body': {'Text': {'Data': body}, 'Html': {'Data': html_body}}
},
Which means that we are sending the contents of email message
twice in one request: once in plain text and once with html tags.
That means our plain text content needs to be much shorter to
make sure we fit within the limit, especially since HTML body
can be much byte-heavier than plain text body.

Hence, we decided to put the limit at 1MB, which is equivalent of
between 250 and 500 pages of text. That's still an extremely long email,
and should be sufficient for all normal use, while at the same
time giving us safe margin while sending the emails through Amazon SES.