Remove single space from alt text of images
This was added as a hack to stop the NVDA screen
reader announcing the full URL of images with
blank alt text.

https://github.com/alphagov/notifications-utils/commit/7879f580625ba70a8b010e5eb96757b20e59d919#diff-5705deaa50ec56f65d19721a2237bbd8b70a5170c47dbea11b4a589586dcd0d4

Testing recently with NVDA and Outlook showed that
images with blank alt text were now announced as
"Graphic".

https://www.pivotaltracker.com/story/show/161183433/comments/219297211

This still doesn't follow any HTML spec but the
experience for screen reader users is now
reasonable enough to be reset to blank.