Fix underscores replaced with asterisks in emails
Emphasis in Markdown can either be applied using underscores or
asterisks.

We format emphasis as the original text, surrounded by asterisks. This
means if you try to use asterisks for emphasis it looks like it’s having
no effect. This is what we want, because we don’t want people using bold
or italic in their emails.

However if you use underscores for emphasis, you get the underscores
replaced by asterisks. This is bad. If, for example, an email address
contains underscores and that email address is quoted in the body of the
email, our code will change that email address to something invalid.

This commit fixes the problem by removing emphasis (and double emphasis)
from the Markdown rules that get applied. This is better than our
previous approach of overriding the rules we didn’t want, where the code
doing the formatting had no knowledge of what characters had been used
to apply the formatting.