Optimise key generation for `Columns` class
`str.translate` seems to be the fastest way of doing character
substitution in Python.

`make_key` is called _a lot_ when doing replacement of placeholders, so
any optimisation we can do on it is a good win.

To test this I temporarily removed the LRU cache on the method, and then
ran the following shell command:
```bash
python -m timeit -v -s "from notifications_utils.columns import Columns" "Columns.make_key('this is _not_ a -test')"
```

Results

Before
```
100000 loops, best of 5: 3.86 usec per loop
```

After
```
200000 loops, best of 5: 1.18 usec per loop
```