Remove strikethrough rule from Markdown parser
Strikethrough in Markdown is specified using double tildes, for example:
```
~~strike~~
```

We don’t support strikethrough, and we remove it by just outputting the
text (for example `strike`).

This is bad if you’re using tildes for other purposes, for example as
characters in a password.

This commit fixes the problem by removing strikethrough from the
Markdown rules that get applied. This is better than our previous
approach of overriding the rules we didn’t want, because it preserves
the text the user has given us, with no overhead.