Remove inline code rule from Markdown parser
Code snippets in Markdown are specified using backticks, for example:
```markdown
`code`
```

We don’t support code snippets, and we remove them by outputting the
text only (for example `code`).

This is bad if you’re using backticks for other purposes, for example
as characters in a password.

This commit fixes the problem by removing inline code snippets from the
Markdown rules that get applied. This is better than our previous
approach of overriding the rules we didn’t want, because it preserves
the text the user has given us, with no overhead.