Fix bug with double semicolons in emails
We use a library called bleach for escaping dangerous HTML in our
template content. It doesn’t seem to handle things which look like HTML
entities (for example `&copy;` or `&gt;`) but contain unexpected
characters (the only one I’ve found so far is a question mark).

This results in the last character of an escape sequence being replaced
with a double semicolon. I haven’t dug into the Bleach source deeply
enough to identify where the problem is, and which other characters
might affect it.

But a fix is to use Python’s builtins to convert any escape sequences to
their unicode equivalent, and then escape them again. This seems to
avoid the same confusion that Bleach encounters.

We can then pass the output of this to Bleach, which passes through
stuff that’s already properly encoded.

This fix is described in this issue comment:

> …
>
> A quick solution that can be implemented by developers is to just
> escape(unescape()) the input before sending it to bleach.
>
> …
>
> This solution does not utilize a HTML parser, it's all regex under
> the hood, so there may be some edge cases. This has passed all our use
> cases though.
>
> While I personally think bleach should handle stuff like this, I don't
> think it is necessary when invoking bleach is only a component of your
> user input sanitization routine.

– https://github.com/mozilla/bleach/issues/453#issuecomment-499182492

There’s one more gotcha, which is that, for historic reasons, we allow
users to include the non-breaking space character (encoded as `&nbsp;`)
in order to create empty paragraphs. This character is ignored by the
Markdown parser unless it’s encoded – just including the unicode
character won’t force an extra paragraph. Python’s `html.escape` doesn’t
encodes things like quotes:
```python
html.escape('"')
'&quot;'
```
But it doesn’t see a need to encoded non-breaking spaces:
```python
html.escape('\xa0')
'\xa0'
```

So to preserve the encoded non-breaking space we also have to
temporarily convert it to something else, do the `unescape`/`escape`
dance, and then convert it back.