Remove workaround for giving bleach HTML-like content
In https://github.com/alphagov/notifications-utils/pull/168 we added a
workaround to stop Bleach trying to interpret template content as HTML
and fix it.

This workaround is now effected by using Python’s native HTML escaping
method to encode the content, so we don’t need to separately encode the
opening character of an HTML tag.