Move polygon processing code from admin app
This is so we can reuse it in the API. We want to make sure that we
always pass the CBC polygons with a sensible point count, no matter
what our clients give us.

There’s basically no unit testing of this code in the admin app, so
this commit adds a bunch of new tests for the various properties.