Stop using bleach for escaping HTML entities
When we first added sanitisation of user-inputted content[1] we chose a
library called Bleach which both escapes and strips HTML from content.

Bleach allows things that are already encoded to be passed through. For
example `G&amp;T` will be outputted as `G&amp;T` and rendered by a
browser as _G&T_. `html.escape` escapes everything, so `G&amp;T` becomes
`G&amp;amp;T` which gets rendered by a browser as _G&amp;T_.

We are relying on Bleach’s behaviour so that one team can pass us
non-breaking spaces as `&nbsp;` to control layout. I also suspect we
will have some teams giving us URLs with `&amp;` already in them, and
seeing _&amp;_ in places you shouldn’t is pretty common (my TV does it
with the programme descriptions on Amazon Prime Video). So it’s good
that we pass this through.

We recently added `html.escape` and `html.unescape` before running
Bleach to work around some buggy handling of things which looked like
entities but aren’t.

This is problematic because `html.unescape` is way too greedy about what
it interprets as an entity. This is because it follows the HTML
specification[3] which, for backwards compatibility, allows some
entities (like `&times;`) to be used without a trailing semicolon (as in
`&times`). This means when someone has a URL which includes
`&timestamp=`, it gets rendered as `&×stamp=` (`&times;` is the entity
for the multiplication symbol `×`, which is subtly different to the
24th letter of the alphabet, `x`).

---

So we can’t:
- use Bleach by itself
- use Bleach in conjunction with `html.unescape`

So rather than trying to find another workaround for Bleach’s buggy
behaviour this commit replaces it with `html.escape` alone, for escaping
content. For stripping HTML from content Bleach is still good.

This commit also adds a tight list of allowed entities which we pass
through by temporarily converting them to unlikely sequences of emoji.
This means we can support `&nbsp;` and `&amp;`, without encountering
unexpected behaviour when someone uses `&timestamp=`.

---

This change performs better than before. When tested with this script:
```bash
python -m timeit -v -s "from notifications_utils.formatters import escape_html; import string; s = (string.ascii_letters + string.punctuation) * 100" "escape_html(s)"
```

Before:
```
50 loops, best of 5: 9.11 msec per loop
```

After
```
10000 loops, best of 5: 37.1 usec per loop (0.0371 msec)
```
245&times; faster

---

1. https://github.com/alphagov/notifications-api/pull/135
2. https://docs.python.org/3/library/html.html#html.escape
3. https://html.spec.whatwg.org/multipage/syntax.html#named-character-references