Support left and right parenthesis escape sequences
The only way to have brackets in the body of a conditional placeholder
is to escape them. Using raw brackets means that Notify will no longer
recognise the placeholder. So we need to explicitly support this
behaviour, and not double-escape or unescape the brackets.

More research in: https://github.com/alphagov/notifications-utils/pull/823#issuecomment-747397880