Improve how we calculate email size
We use this to see if email is too big. The way we counted it
so far was, we used getsizeof method, which includes object size
and garbage collection size. It turns out is works better to encode
the string with utf-8 and then measure it's length with len

Link to relevant discussion on Stack Overflow:
https://stackoverflow.com/questions/4013230/how-many-bytes-does-a-string-have/4013418#4013418