Add code to check length of broadcast templates
We currently don't validate length of broadcast messages, and will get
an error back from the CBCs if we send them a message that is too long.
Instead, we should add that validation when the user is saving the
template on the UI.

There is a defined max length in characters for broadcasts, which is set
at 1,395 character for GSM and 615 for unicode.

Broadcast messages follow the same encoding as SMS (GSM). Characters
from the extended character set should as two characters. If there are
any non-GSM characters we’ll send the whole thing as UCS-2 (unicode).

Adding this code to utils so we can re-use it in the admin and the API.

I’ve also added some useful properties (like `.non_gsm_characters`) that
we can use to write nice error messages.

In the future we might not expose all this complexity to our users, but
it’s good to have it technically correct in this library.