Don’t count unpopulated placeholders
When counting the content of text messages templates we want to be able
to say stuff like ‘2 fragments (not including personalisation)’

This isn’t possible if the names of the placeholders, plus the brackets
are included in the length

We were doing it this way because we weren’t using this count without
personalisation for anything. Now that we will be the count needs to
change to explicitly exclude these characters.

Implementation is slightly more complex because `a ((placeholder)) c`
should count as 4 characters without personalisation. That means that
when it’s replaced with a single character the count is the length of
the content (4 characters for `a  c`) plus the length of the
personalisation (1 character for `b`), so 5 characters for `a b c`. If
we simply replaced the placeholders with empty strings it would be
saying `3 + 1 = 5` (because the content would be counted as `a c`).