Move _get_unsanitised_content to BaseSMSTemplate
This means that any SMS or Broadcast template has access to this method.
Previously `BroadcastMessageTemplate` had access to it (because it
subclassed `SMSMessageTemplate`) but `BroadcastPreviewTemplate` didn’t.
So this commit also reinforces the tests to check character counting
works on both template classes (character counting depends on getting
the unsanitised content, so is where exceptions were being raised).