Add bounds property to Polygons
Knowing the bounds of a polygon lets you make a quick guesses as to
which polygons overlap with other polygons.

An easy way to think about bounds is the imagine a box which contains
a polygon. The bounds are the coordinates of the four points of this
box.

Shapely already has a property for giving the bounds of a polygon. This
commit adds the same property to our `Polygons` class, which gives the
outer bounds of all the polygons combined.