Add method to calculate how much polygons overlap
When we let people draw their own polygons we want to know how similar
these polygons are to other polygons we already know about. If two
polygons are very similar they probably contain about the same number of
mobile phones, for example.

This commit adds methods to do this. It works by generating a new
polygon, representing the intersection (overlap) of two polygons, and
then looking at the area of this new polygon.