Preserve more detail in concave parts of polygons
After analysing the polygon for the Reading test we have decided that
we’re being a bit too aggressive with how much we are smoothing it.

We still need to do a lot of smoothing to group together very disparate
areas like the Western Isles of Scotland.

The limits we’re targeting are:
- Max 6000 points overall (we don’t need anywhere near this detail)
- Max 25 polygons (this is the one that’s harder to hit)

This change increases the maximum number of polygons in a single area
from 6 to 9. That’s still comfortably within the limit of 25. We need
to be comfortable because we can’t be sure what will happen when people
start combining different areas.