Allow Zendesk tickets to be create with internal notes
Sometimes we want to raise a ticket but not let the person raising it
see the auto-generated content of the ticket.

We can acheive this by making the first message on the ticket be an
‘internal note’, rather than a ‘public reply’.

The example of this is where we want to add go live notes to an
organisation, which need to appear in the ticket but might not be
information that we want to divulge to an individual service.