Detect optional placeholders with brackets in their text
This didn't work before.

The new way means we need additional part in regex looking for
opening placeholder brackets that makes sure it also works
with a triple bracket - additional bracket should stay on the
outside of the placeholder.

For regex matching the inside of placeholder brackets, we
use the question mark for a non-greedy match.
Adding ? after the qualifier makes it perform the match
in non-greedy or minimal fashion.
So without it we'd match "((something)))" from
"text with (((something)))" in brackets instead of
the correct "((something))".