Tranform polygons to a Cartesian plane
> Even though the Earth is not flat – and for that matter not exactly
> spherical – there are many analytic problems that can be approached by
> transforming Earth features to a Cartesian plane, applying tried and
> true algorithms, and then transforming the results back to geographic
> coordinates. This practice is as old as the tradition of accurate
> paper maps.

– https://shapely.readthedocs.io/en/stable/manual.html#coordinate-systems

At the moment, we are not doing any transformation of features before
applying geometric algorithms to them. This is, in effect, assuming that
the earth is flat.

This pull request implements the transformation of our polygons to a
Cartesian plane. In other words, it converts them from being defined in
spherical degrees to metres.

This is important because spherical degrees correspond to different
linear distances depending on:
- where on the Earth’s surface you are
- whether you are measuring a north/south or east/west distance

Thus far we have got around this by saying 1˚ is approximately 111km and
pretending that our coordinates are not spherical. However in the north
of Scotland 1˚ measured east/west is only about 50km. So when we say
‘expand a shape by 5km in all directions’ it will only expand by ~2km
in the east/west direction.

Shapely itself does not support coordinate transformations. So we need a
new tool, Pyproj, to convert the spherical coordinates to linear
metres.

It does this using a coordinate reference system, which is a a set of
mathematical rules for specifying how coordinates are to be assigned to
points in a given zone of the earth’s surface. For this commit we
make two naive assumptions:
- that all our polygons are in the [29N zone](https://epsg.io/32629)
- that all polygons need transforming to linear coordinates, no matter
  what, if any operations we are going to do on them