Remove approximate conversions of degrees to linear units
Since we are now doing this conversion properly, with a coordinate
reference system, we no longer need these approximations.

One of the tests still needs to draw a shape of roughly x metres in
length. I’ve kept the approximation here so that we’re not just testing
the self-consistency of the operations.