Use appropriate coordinate reference systems
Previously we were assuming that all our polygons were in the 29N zone.
This is an incorrect assumption because the Great Britain and Northern
Ireland spans 3 different zones.

This commit implements selection of the correct zone based on the bounds
of the provided polygon. This should mean any measurements or operations
we do on these polygons are more accurate.