Don’t lookup CRS if it’s already known
Looking up the coordinate reference system to use is surprisingly
expensive, so we should avoid repeating it wherever possible.

We do this by assuming `Polygons` are in one of two states:
- lists of lists of coordinates with an unknown CRS
- lists of `Polygon` objects with a known CRS

We transform from the former to the latter when needed. If we’re already
in the latter state then we should reuse the CRS that’s already been
looked up.