Refactor polygon smoothing to prevent artefacts
Our polygon smoothing works by expanding and then contracting each
shape.

Because it was using different parameters (resolution and join style)
for the expansion and contraction it was introducing artefacts, i.e. it
was creating entirely new shapes.

This commit refactors the smoothing to re-use our existing `bleed_by`
method to expand and contract the polygon. By re-using this method we
have the same parameters for both the expansion and contraction. This
seems to prevent the artefacts.