Add some extra buffering for small shapes
At the moment our buffering is based on the estimated typical bleed of
a cell broadcast. This works well for larger shapes, but for smaller
shapes it can be a bit destructive, because we end up buffering a lot
relative to the size of the space. This commit reduces the amount
of inward buffer by an additional fixed amount. This helps ensure that
the simplified polygons always contain the original polygons, even for
very small shapes. This should mean that we always ensure the original
polygons are targeted, without increasing the coverage excessively
(because the additional area is only a 15 metres in each direction).