Only transform Polygon objects back to WGS84
Internally the `Polygons` class can either store:
- a list of `Polygon` objects, with a mandatory `utm_crs` code which
  says how to convert them back to degrees
- a list of coordinate lists, with an optional `utm_crs` code which
  speeds up any future conversion to metres (if the code is not provided
  then it has to be derived)

When we output the polygons to display on a map or send to a CBC we need
they must use degrees as their unit. This means that, if they are stored
as `Polygon` objects, we need to first convert them back to degrees. If
they are stored as lists of coordinates then we should just return the
list, even if we also have a `utm_crs` code to use in future conversions
to metres.

In other words the presence of a `utm_crs` code isn’t a good indicator
of whether the polygons are stored in degrees or metres, but the `type`
of the objects is.