Refactor ZendeskClient
When we start creating new Zendesk tickets the data gets a bit more
complicated, so there will be two classes - the `ZendeskClient` to make
the request and the `NotifySupportTicket` to format the data.

This changes the code of the ZendeskClient to make it easier to move
towards the new way of doing things.