Add NotifySupportTicket class to create Zendesk ticket data
At the moment, we have the `ZendeskClient` which formats the data for
the ticket and then sends it to Zendesk. We are going to start using a
custom form in Zendesk, which means that there will be more data and it
will need a bit more formatting. Since this is the case, we will create
a new class to deal with the data formatting and the Zendesk client will
only be responsible for sending data.

We're going to be gradually switching the tickets to use the new form,
so this change introduces the new way of doing things but also keeps the
old. The duplication can be removed once all tickets are created using
the custom form.