Remove `create_ticket` method from ZendeskClient
`create_ticket` used to be used to both create and send a Zendesk
ticket, but this has been replaced by a new way of doing things and is
no longer used in api or admin. This also removes all the constants it
was using since these also exist in the `NotifySupportTicket` class.