Replace non breaking space with regular space
In https://github.com/alphagov/notifications-utils/commit/7a77abedb94c373852916f94f4b95a3eaf93b3ba
the commit message says that we wish to replace a non breaking space
with a regular space. However, the code instead replaces it with no
character at all, so essentially non breaking spaces are removed.

We think this is incorrect and that we do wish to have the behaviour
as described in that commit message as that makes more logical sense
for users.

Therefore, this commit fixes the bug and instead of removing non
breaking spaces, transforms them instead to a regular space.