Bug fix - stop removal of non breaking spaces from templates
Currently, we are

- removing non breaking spaces in email subjects through the use
  of `normalise_whitespace`
- removing non breaking spaces in sms previews through the use of
  `sms_encode`
- removing non breaking spaces from sms sending through the use of
  `normalise_whitespace`

This is not the behaviour we want for our users as it is causing
- either their messages get sent with no space between words
- or they spot there is a mistake and have to put extra spaces in
  to their template to correct for it

We fix this by

- changing `normalise_whitespace` to replace non breaking spaces
  with single spaces and still strip other obscure non width whitespace
- changing `sanitise_sms` to replace non breaking spaces with single spaces
  and also why we are at it, allow non width whitespace to be included
  and replaced with an appropriate empty character

Note, we recognise this is quite a duplicative approach, essentially
we have two places in our code doing the same thing. However we didn't
want to try and deal with the issue of in which order
`normalise_whitespace` and `sms_encode` are called which we are
inconsistent on. Also note `SanitiseText` is also used for deciding
which characters to allow in templates to start with so there are
other knock on effects to consider.

In doing this change, we took the opportunity to rename some
functions for better clarity and more precision.