Remove `normalise_line` which is a duplicate of `normalise_whitespace`
We are currently using `normalise_line` which appears to do
exactly the same as `normalise_whitespace`. The difference is that
normalise whitespace uses a regex instead of using `split` and
`join` to remove whitespace we don't want.

We therefore remove `normalise_line` and replace its use with
`normalise_whitespace`.

This is a breaking change, however it does not appear that
`normalise_line` is being used outside of this repo.