Removes confusing non break space character from unrelated test
This test is to do with newlines and new line characters. Therefore
it is confusing to have a non breaking space character in here as
you may think the test is also testing something around that. I've
removed it so the test is solely about what is in its name.