Bump bleach to 3.3.1 or higher
Pyup is complaining that bleach versions less than 3.11 have a
vulnerability.

I think this is a mistake because Bleach versions go from 3.3.1 to 4.0.0
– there is no version 3.11.

There was a CVE for versions less than 3.3.0:
https://github.com/advisories/GHSA-vv2x-vrpj-qqpq

We shouldn’t be vulnerable here because we’re already on 3.3.0.

Anyway this commit bumps us to a version greater than 4 to stop Pyup
complaining. I’ve specified at least 4.1.0 because that is also the
first version with explicit Python 3.9 support which we’re going to
want soon.