DRY-up Celery boilerplate across our repos
Previously each app had its own version of the class / helper:

- notifications-api: https://github.com/alphagov/notifications-api/blob/master/app/celery/celery.py
- notifications-template-preview: https://github.com/alphagov/notifications-template-preview/blob/master/app/celery/celery.py
- notifications-antivirus: https://github.com/alphagov/notifications-antivirus/blob/master/app/celery/celery.py
- notifications-ftp: https://github.com/alphagov/notifications-ftp/blob/master/app/celery/celery.py
- notifications-govuk-alerts: https://github.com/alphagov/notifications-govuk-alerts/blob/main/app/celery/celery.py

This centralises the code in one place to keep it consistent. It
should be a drop-in replacement in all apps, except:

- notifications-template-preview: need to make the 'celery' config
key uppercase ('CELERY') as that's what we look for here.

- notifications-api: this has an extra 'log_on_worker_shutdown'
hook, which we've agreed to delete as it lacks a solid use case.

- (all apps): we no longer / don't need to set the "main" argument
as we don't run "app/celery/celery.py" as a script (see [2]).

- (all apps): we now set the broker_url as part of the general
"self.conf.update" line, instead of separately.

A final thing to note is that 'typing = False' will now disable
Celery 4+ argument checking across all repos. We should be able to
turn this on once each app is using the new code here, since then
we will have fully rolled out [1]. In the meantime, it's not a big
deal if we don't have it on for a short period.

[1]: https://github.com/alphagov/notifications-api/pull/3364
[2]: https://docs.celeryproject.org/en/stable/reference/celery.html#celery.Celery