Enable Celery argument checking for apply_async
This means we will get an error at the point we invoke an async
task, instead of the task failing silently in the background due
to it being passed incorrect arguments.

Most of our repos already had this feature enabled [1] but for API
we found it interacted badly with our previous "kwargs" approach
to passing around "request_id" state [2].

Since the base class now uses Celery headers to support Request ID
tracing and all repos have been deployed with this for some time,
we can safely remove the code to handle tasks created using the
old "kwargs" code, and enable argument checking.

[1]: https://docs.celeryproject.org/en/stable/userguide/tasks.html#argument-checking
[2]: https://github.com/alphagov/notifications-api/pull/3355/commits/3ecbdbb2609a9edcca5be1892053d8379bb6b580