Let users of request cache specify a custom TTL
We have a case where data in our cache is becoming stale. Normally we
would update the cached value as soon as it becomes stale but this is
not possible because the data is populated asynchronously.

Instead we want to reduce the amount of time that the stale data is
sitting around for.

At the moment RequestCache keeps data for 7 days (in line with our
default data retention policy) and this is hard coded.

This pull request adds a keyword argument to RequestCache.set, allowing
any of its user to specify their own cache expiry time.

Units in seconds for consistency with the underlying library.

Note that Redis requires integer values for its TTL argument.