Timeout processing RecipientCSV rows
Processing a large CSV with this code can take a loooooong time.
Although we've tried to improve the processing time [1] and put
in checks before we run this code [2], it's possible there could
be certain CSVs that still take a long time to process, or that
the performance improvements don't last.

Regarding the default timeout, analysis of request times for the
past month show the majority are within 10s:

    0-10s: 99.86%
    10-15s: 0.02%
    15-20s: 0.01%
    20-25s: 0.00%

However, analysis of "reasonably large" CSVs shows that we still
need a high timeout for those cases:

    Email CSV (1 column): average of 6.5s
    Email CSV (4 columns): average of 10s
    Phone CSV (4 columns): average of 13s
    Phone CSV (4 columns): average of 15s (with 32K errors for invalid numbers)
    Phone CSV (8 columns): average of 20s (with 32K errors)

Equally, we don't want to keep the user waiting while _we_ do an
action. We decided an overridable default of 20s is acceptable.

In terms of implementation, while there are packages that do this,
I don't think it's worth installing one for the small number of
lines it takes to implement this - we still need to test it as well.

[1]: #930
[2]: alphagov/notifications-admin#4090