Revert "Timeout processing RecipientCSV rows"
This reverts commit 4fdf28412c4a74824f06607a19e9893b9e730326.

It turns out showing a contextual message for this error is hard,
as we would need to either:

- Show the error on an existing preview page, but be very careful
to structure the code so that none of the other checks run, since
these would trigger more timeout errors.

- Introduce a new error page. This is unprecedented for previewing
a CSV, where we always try to show a preview - this isn't possible
if the CSV is taking too long to process.

Taking a step back, we shouldn't really be adding timeout wrappers
around our code. A more future-proof approach would be:

- To monitor timings for all requests to our apps.
- Take action to improve performance where necessary.

While it's still possible to get this code to run for a long time,
the improvements we've already made should reduce the likelihood
of this being an issue in practice.