Log stacktrace when a task fails
This should help debug the failures we're seeing for one of the
nightly tasks - create-nightly-notification-status-for-day - even
though the task gets to the point of logging a success.

It's worth noting Celery is probably emitting useful logs, but we
supress them [1]. Originally this used to be an exception log [2],
but somehow it got changed to error when we centralised it [3].
Changing to an exception log will mean we get the stacktrace as
part of the log message [4] automatically.

The call to "super()" is unnecessary:

- We don't do it for "on_success".
- It wasn't tested.
- The overridden method does nothing [5].
- None of the docs say to call "super()".

[1]: https://github.com/alphagov/notifications-api/blob/master/scripts/paas_app_wrapper.sh#L9
[2]: https://github.com/alphagov/notifications-api/commit/2622866622cb331f79e208dcee389bca480012f1#diff-76936416943346b5f691dac57a64acebc6a1227293820d1d9af4791087c9fb9eR10
[3]: https://github.com/alphagov/notifications-utils/commit/8765f84a613e5bc0036daf2f036cc950308810f3#diff-9f2919d5016ba611c744a546be5484fc79b037248a6a692dc14befe2534912f8R57
[4]: https://docs.python.org/3/library/logging.html#logging.Logger.debug
[5]: https://github.com/celery/celery/blob/5b698151d5e8da10f6706df42fb99fb3105ac025/celery/app/task.py#L1024