Rename columns module to ‘insensitive dict’
Originally the `Columns` class was just used for representing the
columns in a spreadsheet row.

However its now used more generally when we need a dictionary whose keys
are not sensitive to case or whitespace. So this commit is the first
part of renaming it to `InsensitiveDict`. This commit renames the
module, but not the class.

I’ve gone with the name `InsensitiveDict` because it follows the pattern
set by `OrderedDict`, a popular Python package.