Move `Row` and `Cell` classes to recipients module
The `Row` and `Cell` classes are used only by the parts of the code
which handle spreadsheets uploaded by our users. So they sit more
naturally in the same module as the main `RecipientCSV` class, not in
the `insensitive_dict` module which has been renamed to reflect its
more general usage.