Remove validate_recipient function
This function only switches between two template types, so the `dict`
lookup is overkill (it probably made a bit more sense when the function
accepted three template types).

Since it’s only called from one place we can inline it and replace it
with a couple of `if` statements. This is fewer lines of code, and less
indirection.