Unify behaviour of Redis delete functions
Previously the main "delete" function would just log exceptions,
whereas the -by-pattern version would let them bubble up. This is
surprising and we think it's an oversight when -by-pattern was
introduced [1]. In particular, if Redis goes down any usages of
this method will lead to 500s, unlike all the others.

We debated which behaviour is better:

- Letting errors propagate will help prevent any updates while Redis
is down, which means the data in it is still valid when it returns.
This prevents a message being sent with a stale template [2].

- Logging errors means users can continue using Notify transparently
to the outage, which is important for things like broadcasts. It does
mean we need to rush to clear the cache when it returns.

For the time being, we think it's worth at least making the behaviour
consistent, so here we go with the majority (log only).

[1]: https://github.com/alphagov/notifications-utils/commit/8cb43fc145c70e0ba838b8bedfdc20538b0cccc2
[2]: https://github.com/alphagov/notifications-admin/commit/01a3df6edce08e3cf0cca42b60dcf95127155497#diff-9a4bee453dcac10f7fa51a00a0f5e25db3947c68bacbc898c9d0f7bad00109c7R216