Unpin all dependencies and remove redundant ones
We decided we prefer to use unpinned dependencies in this repo [^1]
so this makes all of them consistent.

I've also removed two dependencies which were added in response to
PyUp generating to many update PRs [^2]. Since we now have a strategy
I think we should put them back and solve the issue more explicitly
in the apps that need them.

[^1]: https://docs.google.com/document/d/1DO6nFV9E7Qr94_NnUfYsMKFrKTvcjn0ESjjDHodfQUk/edit#heading=h.pohfahyz07ro
[^2]: https://github.com/alphagov/notifications-utils/commit/d0fa02026d525ea537785b91255db182b01e6caf