Try upgrading mistune to 2.0.2
As far as I can see there's no upgrade path from 0.8.4 that doesn't
involve a major overhaul of our use of this library, which is only
in formatters.py (not in other repos).

Some of the differences to overcome:

- The inbuilt rules are harder to override or don't seem to have an
equivalent - lists in particular.

- Previously togglable features like "emphasis" now need disabling
with a hack (a regex that never matches).

- Many of the rendered method signatures have changed in sublte ways.

This commit gets many of the tests passing but not all of them. On
the whole it's unclear if the new version of the library gives us any
advantage over the previous one.

Pragmatic suggestion: consider v2 of the library to be a compeletely
separate library and never upgrade it.