Pin mistune below totally incompatible version 2
Although the main, public behaviour is similar, we rely on changing
the internals of mistune, which are completely different in v2.

Example attempt at upgrading:
https://github.com/alphagov/notifications-utils/compare/try-upgrade-mistune?expand=1

The version of mistune we're using doesn't depend on anything [^1],
so there's no risk of it locking versions of other dependencies.

The only risk I can think of is that something else we use now or in
future will require a newer version of this library. If we get to that
point a pragmatic option would be to just vendor the code we're using
now, which was all in a single file prior to v2 [^2].

[^1]: https://github.com/lepture/mistune/blob/6c467d7333744409dea9206d2d58957be171d156/setup.py
[^2]: https://github.com/lepture/mistune/blob/6c467d7333744409dea9206d2d58957be171d156/mistune.py